* commit 'c35e13382e9ee4cdeaa4d501f1a0b834fd6b0e8e': (519 commits)
  Cropping footstep planner output to match ros message capacity when being published
  Removed a bunch of keep SCS true.
  :arrow-up: ihmc-graphics-description 0.19.4
  :arrow_up: Upgrade IHMC libraries to latest version.
  :arrow_up: ihmc-cd 1.20
  Removed valvenet and cuda and added a comment to ValveDetector.
  Missed usages of text-to-speech.
  Removed the text-to-speech code.
  Fix ValveNet.
  :arrow_up: ihmc-build 0.25.0 Remove Bintray declaration.
  :arrow_up: ihmc-jmonkey-engine-toolkit 0.19.7
  Switch to newly published version of JME on Maven Central.
  Fix build. Use us.ihmc group for jinput.
  Add old script that's working better.
  Fix ihmc-ros-control.
  Add interactive arg to fix interaction.
  Rename Multisense Docker image to multisense-ros and switch to using a pre cloned workspace.
  default experimental physics engine to true but allow for it to be changed
  Add Multisense image, enable passwordless sudo of robotlab user, and add DNS and network debugging tools.
  Implement saving and loading environments with JSON.
  ...

+++ ihmc-ros-tools/src/main/java/us/ihmc/utilities/ros/publisher/RosMapsenseConfigurationPublisher.java
@@ -4,13 +4,12 @@ import map_sense.MapsenseConfiguration;
 
 public class RosMapsenseConfigurationPublisher extends RosTopicPublisher<MapsenseConfiguration>
 {
-
    public RosMapsenseConfigurationPublisher()
    {
       super(MapsenseConfiguration._TYPE, false);
    }
 
-   public void publish(short kernelLevel, short filterSize, float mergeAngular, float mergeDistance)
+   public void publish(byte kernelLevel, byte filterSize, float mergeAngular, float mergeDistance)
    {
       MapsenseConfiguration message = getMessage();
       message.setKernelLevel(kernelLevel);
@@ -19,5 +18,4 @@ public class RosMapsenseConfigurationPublisher extends RosTopicPublisher<Mapsens
       message.setMergeDistanceThreshold(mergeDistance);
       publish(message);
    }
-
 }
